+++ mne/__init__.py
@@ -43,7 +43,7 @@ from .cov import (read_cov, write_cov, Covariance, compute_raw_covariance,
 from .event import (read_events, write_events, find_events, merge_events,
                     pick_events, make_fixed_length_events, concatenate_events,
                     find_stim_steps, AcqParserFIF)
-from ._freesurfer import (head_to_mni, head_to_mri, read_talxfm,
+from ._freesurfer import (get_subjects_dir, head_to_mni, head_to_mri, read_talxfm,
                           get_volume_labels_from_aseg, read_freesurfer_lut)
 from .forward import (read_forward_solution, apply_forward, apply_forward_raw,
                       average_forward_solutions, Forward,
