+++ mne/viz/_3d.py
@@ -21,6 +21,7 @@ import numpy as np
 
 from ..defaults import DEFAULTS
 from ..fixes import _crop_colorbar, _get_img_fdata, _get_args
+from .._freesurfer import get_subjects_dir
 from ..io import _loc_to_coil_trans
 from ..io.pick import pick_types, _picks_to_idx
 from ..io.constants import FIFF
@@ -35,7 +36,7 @@ from ..transforms import (_find_trans, apply_trans, rot_to_quat,
                           combine_transforms, _get_trans, _ensure_trans,
                           invert_transform, Transform, rotation,
                           read_ras_mni_t, _print_coord_trans)
-from ..utils import (get_subjects_dir, logger, _check_subject, verbose, warn,
+from ..utils import (logger, _check_subject, verbose, warn,
                      has_nibabel, check_version, fill_doc, _pl, get_config,
                      _ensure_int, _validate_type, _check_option)
 from .utils import (mne_analyze_colormap, _get_color_list,
